+++ zhmcclient/_activation_profile.py
@@ -78,7 +78,22 @@ class ActivationProfileManager(BaseManager):
         #     * `reset`: Reset Activation Profiles
         #     * `image`: Image Activation Profiles
         #     * `load`: Load Activation Profiles
-        super(ActivationProfileManager, self).__init__(ActivationProfile, cpc)
+
+        # Resource properties that are supported as filter query parameters.
+        # If the support for a resource property changes within the set of HMC
+        # versions that support this type of resource, this list must be set up
+        # for the version of the HMC this session is connected to.
+        query_props = [
+            'name',
+        ]
+
+        super(ActivationProfileManager, self).__init__(
+            resource_class=ActivationProfile,
+            parent=cpc,
+            uri_prop='element-uri',
+            name_prop='name',
+            query_props=query_props)
+
         self._profile_type = profile_type
 
     @property
@@ -101,7 +116,7 @@ class ActivationProfileManager(BaseManager):
         return self._profile_type
 
     @_log_call
-    def list(self, full_properties=False):
+    def list(self, full_properties=False, filter_args=None):
         """
         List the Activation Profiles of the type managed by this object and in
         this CPC.
@@ -112,6 +127,14 @@ class ActivationProfileManager(BaseManager):
             Controls whether the full set of resource properties should be
             retrieved, vs. only the short set as returned by the list
             operation.
+
+          filter_args (dict):
+            Filter arguments that narrow the list of returned resources to
+            those that match the specified filter arguments. For details, see
+            :ref:`Filtering`.
+
+            `None` causes no filtering to happen, i.e. all resources are
+            returned.
 
         Returns:
 
@@ -124,19 +147,29 @@ class ActivationProfileManager(BaseManager):
           :exc:`~zhmcclient.AuthError`
           :exc:`~zhmcclient.ConnectionError`
         """
-        activation_profiles_name = self._profile_type + '-activation-profiles'
-        profiles_res = self.session.get(self.cpc.uri + '/' +
-                                        activation_profiles_name)
-        profile_list = []
-        if profiles_res:
-            profile_items = profiles_res[activation_profiles_name]
-            for profile_props in profile_items:
-                profile = ActivationProfile(self, profile_props['element-uri'],
-                                            None, profile_props)
-                if full_properties:
-                    profile.pull_full_properties()
-                profile_list.append(profile)
-        return profile_list
+        query_parms, client_filters = self._divide_filter_args(filter_args)
+
+        resources_name = self._profile_type + '-activation-profiles'
+        uri = '{}/{}{}'.format(self.cpc.uri, resources_name, query_parms)
+
+        resource_obj_list = []
+        result = self.session.get(uri)
+        if result:
+            props_list = result[resources_name]
+            for props in props_list:
+
+                resource_obj = self.resource_class(
+                    manager=self,
+                    uri=props[self._uri_prop],
+                    name=props.get(self._name_prop, None),
+                    properties=props)
+
+                if self._matches_filters(resource_obj, client_filters):
+                    resource_obj_list.append(resource_obj)
+                    if full_properties:
+                        resource_obj.pull_full_properties()
+
+        return resource_obj_list
 
 
 class ActivationProfile(BaseResource):
@@ -166,9 +199,7 @@ class ActivationProfile(BaseResource):
             raise AssertionError("ActivationProfile init: Expected manager "
                                  "type %s, got %s" %
                                  (ActivationProfileManager, type(manager)))
-        super(ActivationProfile, self).__init__(manager, uri, name, properties,
-                                                uri_prop='element-uri',
-                                                name_prop='name')
+        super(ActivationProfile, self).__init__(manager, uri, name, properties)
 
     def update_properties(self, properties):
         """