+++ zhmcclient/_nic.py
@@ -50,7 +50,13 @@ class NicManager(BaseManager):
         # Parameters:
         #   partition (:class:`~zhmcclient.Partition`):
         #     Partition defining the scope for this manager.
-        super(NicManager, self).__init__(Nic, partition)
+
+        super(NicManager, self).__init__(
+            resource_class=Nic,
+            parent=partition,
+            uri_prop='element-uri',
+            name_prop='name',
+            query_props=[])
 
     @property
     def partition(self):
@@ -60,7 +66,7 @@ class NicManager(BaseManager):
         """
         return self._parent
 
-    def list(self, full_properties=False):
+    def list(self, full_properties=False, filter_args=None):
         """
         List the NICs in this Partition.
 
@@ -70,6 +76,14 @@ class NicManager(BaseManager):
             Controls whether the full set of resource properties should be
             retrieved, vs. only the short set as returned by the list
             operation.
+
+          filter_args (dict):
+            Filter arguments that narrow the list of returned resources to
+            those that match the specified filter arguments. For details, see
+            :ref:`Filtering`.
+
+            `None` causes no filtering to happen, i.e. all resources are
+            returned.
 
         Returns:
 
@@ -82,15 +96,22 @@ class NicManager(BaseManager):
           :exc:`~zhmcclient.AuthError`
           :exc:`~zhmcclient.ConnectionError`
         """
-        nic_uris = self.partition.get_property('nic-uris')
-        nic_list = []
-        if nic_uris:
-            for uri in nic_uris:
-                nic = Nic(self, uri)
-                if full_properties:
-                    nic.pull_full_properties()
-                nic_list.append(nic)
-        return nic_list
+        resource_obj_list = []
+        uris = self.partition.get_property('nic-uris')
+        if uris:
+            for uri in uris:
+
+                resource_obj = self.resource_class(
+                    manager=self,
+                    uri=uri,
+                    name=None,
+                    properties=None)
+
+                if self._matches_filters(resource_obj, filter_args):
+                    resource_obj_list.append(resource_obj)
+                    if full_properties:
+                        resource_obj.pull_full_properties()
+        return resource_obj_list
 
     def create(self, properties):
         """
@@ -189,9 +210,7 @@ class Nic(BaseResource):
         if not isinstance(manager, NicManager):
             raise AssertionError("Nic init: Expected manager type %s, got %s" %
                                  (NicManager, type(manager)))
-        super(Nic, self).__init__(manager, uri, name, properties,
-                                  uri_prop='element-uri',
-                                  name_prop='name')
+        super(Nic, self).__init__(manager, uri, name, properties)
 
     def delete(self):
         """